# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: MarkFlow API
  version: 1.1.0-draft
  description: 'MarkFlow — 마크다운 노드 기반 실시간 협업 캔버스 API 명세. REST는 이 문서가 기계 판독 정본이다. 실시간(멀티커서·노드
    동기화·소프트 락·채팅)은 Socket.io 직접 구현이 정본이며, 이벤트/페이로드 계약 정본은 packages/shared/src(SOCKET_EVENTS +
    zod schema)다. v1.1은 ActivityLog, 노드 휴지통·영구 삭제 엔드포인트를 반영한다(프로젝트는 하드 삭제 — 휴지통 없음).'
servers:
- url: https://api.markflow.app/api
  description: Production example
- url: http://localhost:3000/api
  description: Local development
...
security: &id001
- bearerAuth: []
paths:
  /auth/signup:
    post:
      tags:
      - Auth
      summary: 회원가입
      operationId: signup
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '201':
          description: Created — body에 accessToken, Set-Cookie로 refresh_token(httpOnly)
          headers:
            Set-Cookie:
              description: refresh_token=<token>; HttpOnly; SameSite=Lax; Path=/
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: ValidationError
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags:
      - Auth
      summary: 로그인
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: OK — body에 accessToken, Set-Cookie로 refresh_token(httpOnly)
          headers:
            Set-Cookie:
              description: refresh_token=<token>; HttpOnly; SameSite=Lax; Path=/
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/me:
    get:
      tags:
      - Auth
      summary: 내 정보 조회
      operationId: getMe
      security: *id001
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags:
      - Auth
      summary: 토큰 갱신 (쿠키의 refresh_token 사용, Bearer 불필요)
      operationId: refreshToken
      security: []
      description: Cookie의 refresh_token으로 인증. 성공 시 새 accessToken(body) + 새 refresh_token(Set-Cookie). 토큰 회전 적용.
      responses:
        '200':
          description: OK — 새 accessToken(body) + Set-Cookie로 새 refresh_token
          headers:
            Set-Cookie:
              description: refresh_token=<new_token>; HttpOnly; SameSite=Lax; Path=/
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
        '401':
          description: Unauthorized — refresh_token 없음·만료·무효
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      tags:
      - Auth
      summary: 로그아웃 (쿠키의 refresh_token 무효화)
      operationId: logout
      security: *id001
      description: DB에서 refresh_token 삭제 + Set-Cookie로 쿠키 제거. refresh_token 없어도 200 반환.
      responses:
        '200':
          description: OK — 쿠키 제거
          headers:
            Set-Cookie:
              description: refresh_token=; expires=Thu, 01 Jan 1970 00:00:00 GMT
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          properties:
            code:
              type: string
              enum:
              - VALIDATION_ERROR
              - UNAUTHORIZED
              - FORBIDDEN
              - NOT_FOUND
              - CONFLICT
              - UNPROCESSABLE
              - INTERNAL
              example: FORBIDDEN
            message:
              type: string
              example: 이 작업을 수행할 권한이 없습니다.
            details:
              type:
              - object
              - array
              - string
              - 'null'
              example: null
    User:
      type: object
      required:
      - id
      - email
      - name
      properties:
        id: *id003
        email:
          type: string
          format: email
          example: user@markflow.app
        name:
          type: string
          example: 임민규
    SignupRequest:
      type: object
      required:
      - name
      - email
      - password
      properties:
        name:
          type: string
          example: 임민규
        email:
          type: string
          format: email
          example: user@markflow.app
        password:
          type: string
          minLength: 8
          format: password
          example: secret123
    LoginRequest:
      type: object
      required:
      - email
      - password
      properties:
        email:
          type: string
          format: email
          example: user@markflow.app
        password:
          type: string
          format: password
          example: secret123
    AuthResponse:
      type: object
      required:
      - accessToken
      - user
      properties:
        accessToken:
          type: string
          example: eyJhbGci...
        user:
          $ref: '#/components/schemas/User'
    RefreshResponse:
      type: object
      required:
      - accessToken
      properties:
        accessToken:
          type: string
          example: eyJhbGci...
...